requirements.txt
gTTS
aiogram
python-dotenv